Fix missing save-image icon by removing MANIFEST
"Save As Image" ToolButton icon was missing, but only under Debian and
Ubuntu, and only when the activity was installed from a package, not
when installed from a bundle.

Debian packaging with cdbs uses MANIFEST if it is present.
(see cdbs:/usr/share/cdbs/1/class/python-sugar.mk)

MANIFEST was deprecated a long time ago, and is not used by
sugar-toolkit or sugar-toolkit-gtk3.

10ea0ef97f1301daed0d7766c04780a134ca0570 added save-image.svg without
adding to MANIFEST.

Rather than add to MANIFEST, this patch removes MANIFEST.